test: ipv4 and ipv6 validation
