+++ substrate-node/support/src/lib.rs
@@ -1,6 +1,9 @@
 #![cfg_attr(not(feature = "std"), no_std)]
 
+pub mod constants;
+pub mod resources;
 pub mod traits;
 pub mod types;
-pub mod resources;
-pub mod constants;
+
+#[cfg(test)]
+mod tests;
