+++ substrate-node/support/src/resources.rs
@@ -114,150 +114,3 @@ impl Resources {
         return self.hru == 0 && self.sru == 0 && self.cru == 0 && self.mru == 0;
     }
 }
-
-#[cfg(test)]
-mod test {
-    use super::*;
-
-    #[test]
-    fn test_calc_cu_falsy_values() {
-        let resources = Resources {
-            hru: 0,
-            cru: 0,
-            mru: 0,
-            sru: 0,
-        };
-
-        let cu = resources.get_cu();
-        assert_eq!(cu, 0);
-    }
-
-    #[test]
-    fn test_calc_cu() {
-        let resources = Resources {
-            hru: 4 * GIGABYTE as u64 * 1024,
-            cru: 64,
-            mru: 64 * GIGABYTE as u64 * 1024,
-            sru: 12 * GIGABYTE as u64 * 1024,
-        };
-
-        let cu = resources.get_cu();
-        assert_eq!(cu, 256);
-    }
-
-    #[test]
-    fn test_calc_cu_2() {
-        let resources = Resources {
-            hru: 4 * GIGABYTE as u64 * 1024,
-            cru: 4,
-            mru: 8,
-            sru: 12 * GIGABYTE as u64 * 1024,
-        };
-
-        let cu = resources.get_cu();
-        assert_eq!(cu, 2);
-    }
-
-    #[test]
-    fn test_calc_su() {
-        let resources = Resources {
-            hru: 4 * GIGABYTE as u64 * 1024,
-            cru: 64,
-            mru: 64,
-            sru: 12 * GIGABYTE as u64 * 1024,
-        };
-
-        let su = resources.get_su();
-        assert_eq!(su, 52);
-    }
-
-    #[test]
-    fn test_calc_su_2() {
-        let resources = Resources {
-            hru: 0,
-            cru: 64,
-            mru: 64,
-            sru: 12 * GIGABYTE as u64 * 1024,
-        };
-
-        let su = resources.get_su();
-        assert_eq!(su, 49);
-    }
-
-    #[test]
-    fn test_calc_su_3() {
-        let resources = Resources {
-            hru: 0,
-            cru: 64,
-            mru: 64,
-            sru: 0,
-        };
-
-        let su = resources.get_su();
-        assert_eq!(su, 0);
-    }
-
-    #[test]
-    fn test_calc_su_4() {
-        let resources = Resources {
-            hru: 4 * GIGABYTE as u64 * 1024,
-            cru: 64,
-            mru: 64,
-            sru: 0,
-        };
-
-        let su = resources.get_su();
-        assert_eq!(su, 3);
-    }
-
-    #[test]
-    fn test_resources_diff() {
-        let resources = Resources {
-            hru: 4 * GIGABYTE as u64 * 1024,
-            cru: 64,
-            mru: 64 * GIGABYTE as u64,
-            sru: 0,
-        };
-
-        let new_resources = Resources {
-            hru: 4 * GIGABYTE as u64 * 1024,
-            cru: 64,
-            mru: 64 * GIGABYTE as u64,
-            sru: 0,
-        };
-
-        assert_eq!(Resources::has_changed(&resources, &new_resources, 1), false);
-
-        let resources = Resources {
-            hru: 4 * GIGABYTE as u64 * 1024,
-            cru: 64,
-            mru: 64 * GIGABYTE as u64,
-            sru: 0,
-        };
-
-        let new_resources = Resources {
-            hru: 4 * GIGABYTE as u64 * 1024,
-            cru: 64,
-            mru: 40 * GIGABYTE as u64,
-            sru: 0,
-        };
-
-        assert_eq!(Resources::has_changed(&resources, &new_resources, 1), true);
-
-        let resources = Resources {
-            hru: 4 * GIGABYTE as u64 * 1024,
-            cru: 64,
-            mru: 64 * GIGABYTE as u64,
-            sru: 1000 * GIGABYTE as u64,
-        };
-
-        let new_resources = Resources {
-            hru: 4 * GIGABYTE as u64 * 1024,
-            cru: 64,
-            mru: 64 * GIGABYTE as u64,
-            sru: 989 * GIGABYTE as u64,
-        };
-
-        assert_eq!(Resources::has_changed(&resources, &new_resources, 1), true);
-    }
-}
